requirements.txt
anyio==3.6.1
asgiref==3.5.2
asyncpg==0.26.0
bcrypt==4.0.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
databases==0.6.1
dnspython==2.2.1
docutils==0.19
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.82.0
flit==3.7.1
flit_core==3.7.1
greenlet==1.1.3
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
orjson==3.8.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.1
PyJWT==2.4.0
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.40
starlette==0.19.1
tomli==2.0.1
tomli_w==1.0.0
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.12
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.3
